pyproject.toml
[tool.poetry]
name = "tms_diploma"
version = "0.1.0"
description = "TMS Diploma"
authors = ["Ihar Shauchuk <shauchuk.ihar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
python-decouple = "^3.6"
flake8 = "^4.0.1"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^6.1.0"}
djangorestframework = "^3.13.1"
Pillow = "^9.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
